requirements.txt
absl-py==0.9.0
astor==0.8.1
attrs==19.3.0
bert-serving-client==1.10.0
bert-serving-server==1.10.0
gast==0.3.3
google-pasta==0.2.0
GPUtil==1.4.0
grpcio==1.27.2
h5py==2.10.0
importlib-metadata==1.5.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
Markdown==3.2.1
more-itertools==8.2.0
numpy==1.18.2
opencc-python-reimplemented==0.1.5
packaging==20.3
pluggy==0.13.1
protobuf==3.11.3
py==1.8.1
pyparsing==2.4.6
pytest==5.3.5
pyzmq==19.0.0
scipy==1.1.0
six==1.14.0
tensorboard==1.14.0
tensorflow==1.14.0
tensorflow-estimator==1.14.0
termcolor==1.1.0
wcwidth==0.1.9
Werkzeug==1.0.0
wrapt==1.12.1
zipp==3.1.0
